chore: more review comments
